+++ src/extension/src/styleChoices.ts
@@ -1,5 +1,5 @@
-import { codeBlockThemeOptions, colorOptions, fontFamilyOptions, fontSizeOptions, legendOptions } from '../../config/style'
-import { themeOptions } from '../../config/theme'
+import { codeBlockThemeOptions, colorOptions, fontFamilyOptions, fontSizeOptions, legendOptions } from '@/config/style'
+import { themeOptions } from '@/config/theme'
 
 export {
   codeBlockThemeOptions,
