+++ src/utils/MDAlert.ts
@@ -1,6 +1,6 @@
-import type { MarkedExtension, Tokens } from 'marked'
 import type { AlertOptions, AlertVariantItem } from '@/types'
-import { getStyleString } from './css-helper'
+import type { MarkedExtension, Tokens } from 'marked'
+import { getStyleString } from '.'
 
 /**
  * https://github.com/bent10/marked-extensions/tree/main/packages/alert
